feat: moves page content from metadata to page
The page content was previously located in the metadata file.
This commit moves the page content to the page file.
The metadata file now only contains the metadata for the page.

This improves the organization of the codebase.
